splitted NativeHttpClient
