+++ log/testomatio.log.1
@@ -398,3 +398,83 @@ SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.Propert
 SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
 ????. 29, 2025 8:19:23 ?? com.testomatio.reporter.core.GlobalRunManager lambda$registerShutdownHook$0
 INFO: JVM is shutting down, finalizing test run...
+????. 29, 2025 10:05:02 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:04 ?? com.testomatio.reporter.client.http.NativeHttpClient createUri
+SEVERE: Invalid URI syntax for request
+????. 29, 2025 10:05:38 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.client.TestomatApiClient finishTestRun
+SEVERE: Failed to finish test run with uid: test-run-uid-123
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.client.TestomatApiClient createRun
+INFO: Testomat.io java reporter version: 0.1.0
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.client.TestomatApiClient createRun
+INFO: See run aggregation at: null
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.client.TestomatApiClient reportTests
+SEVERE: Failed to report batch test /nNetwork error
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.client.TestomatApiClient createRun
+INFO: Testomat.io java reporter version: 0.1.0
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.client.TestomatApiClient createRun
+INFO: See run aggregation at: null
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:39 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:40 ?? com.testomatio.reporter.core.batch.BatchResultManager sendBatch
+SEVERE: Failed to report batch (attempt 3/3): Network error
+????. 29, 2025 10:05:40 ?? com.testomatio.reporter.core.batch.BatchResultManager sendBatch
+SEVERE: Failed to report 1 tests after 3 attempts
+????. 29, 2025 10:05:40 ?? com.testomatio.reporter.core.batch.BatchResultManager shutdown
+WARNING: BatchResultManager shutdown with 1 failed results
+????. 29, 2025 10:05:40 ?? com.testomatio.reporter.core.batch.BatchResultManager addResult
+WARNING: BatchResultManager is not active, skipping result: Test 1
+????. 29, 2025 10:05:40 ?? com.testomatio.reporter.core.batch.BatchResultManager sendBatch
+SEVERE: Failed to report batch (attempt 1/3): Network error
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.batch.BatchResultManager shutdown
+WARNING: BatchResultManager shutdown with 1 failed results
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.batch.BatchResultManager shutdown
+WARNING: BatchResultManager shutdown with 1 failed results
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:42 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:43 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:43 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:43 ?? com.testomatio.reporter.core.framework_integration.JUnitExtension handleTestResult
+WARNING: No test method found in context, cannot report test result
+????. 29, 2025 10:05:43 ?? com.testomatio.reporter.core.framework_integration.JUnitExtension handleTestResult
+WARNING: No test method found in context, cannot report test result
+????. 29, 2025 10:05:43 ?? com.testomatio.reporter.core.framework_integration.JUnitExtension handleTestResult
+WARNING: No test method found in context, cannot report test result
+????. 29, 2025 10:05:43 ?? com.testomatio.reporter.core.framework_integration.JUnitExtension handleTestResult
+WARNING: No test method found in context, cannot report test result
+????. 29, 2025 10:05:43 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager finalizeRun
+INFO: Test run finished: test-uid-123
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager finalizeRun
+SEVERE: Failed to finish test runnull
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager finalizeRun
+INFO: Test run finished: test-uid-123
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager initializeIfNeeded
+SEVERE: Failed to initialize test run: com.testomatio.reporter.exception.PropertyNotFoundException: No such property: testomatio.api.key
+????. 29, 2025 10:05:44 ?? com.testomatio.reporter.core.GlobalRunManager lambda$registerShutdownHook$0
+INFO: JVM is shutting down, finalizing test run...
