+++ src/main/java/com/testomatio/reporter/client/http/retryable/RetryableRequestExecutor.java
@@ -1,0 +1,9 @@
+package com.testomatio.reporter.client.http.retryable;
+
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+
+public interface RetryableRequestExecutor {
+    HttpResponse<String> executeRetryable(HttpRequest request, HttpClient client);
+}
